requirements.txt
Flask==3.0.2
discord.py==2.3.2
python-dotenv==1.0.1
sqlalchemy==2.0.25
flask_sqlalchemy==3.1.1
psycopg2-binary==2.9.9
tabulate==0.9.0
